requirements.txt
numpy~=1.19.3
opencv-python~=3.4.5.20
matplotlib~=3.3.2
Keras~=2.2.4
h5py==2.10
Pillow~=8.1.2
scipy~=1.5.4
tensorflow~=1.15.0
albumentations~=0.5.2
scikit-image~=0.17.2